pyproject.toml
[tool.poetry]
name = "cps-info-viz"
version = "0.1.0"
description = ""
authors = ["Louie Lu <git@louie.lu>"]
readme = "README.md"
packages = [{include = "cps_info_viz"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.1"
pycpsdata = "^1.0.1"
numpy = "^1.26.4"
matplotlib = "^3.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
